Added functionaility to rturn stock
there's a bug here I can't figure out, despite all indications and debugging the DbContext won't seem to update, so for the sake of time I'm moving on. I suspect the issue may stem from my unfamiliarity with in memory C# Dbs as used in this repo. Rather than spending hours debugging I'm going to implement the rest of the requested features