Add functionality to levy fines for late returns
this functionality required midifying the borrower class to add a property to track the total fines owed. I then updated the borrower repository to enable updating of this field in memory. I then expanded on my loan service to levy these fines against the borrower as part of the return process. I kept a simple fine model for the purpose of this test however this would be ideal for pulling into another method and abstracting which would enable dynamic pricing and other changng models to be implemented